# docs/swagger.yaml
basePath: /v1
definitions:
  controllers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.LoginDto:
    properties:
      login:
        description: Login using username, e-mail, or phone.
        example: Login
        type: string
      password:
        description: |-
          Password must contain at least one uppercase letter, one lowercase letter, one digit, and one special symbol.
          It can consist of printable ASCII characters.
        example: 8p6R*R{3
        maxLength: 72
        minLength: 8
        type: string
    required:
    - login
    - password
    type: object
...
  dto.UpdatePasswordDto:
    description: The 'NewPassword' and 'OldPassword' should not match, while the 'ConfirmPassword'
      is required for verification but must match the new password.
    properties:
      confirmPassword:
        description: |-
          Password must contain at least one uppercase letter, one lowercase letter, one digit, and one special symbol.
          It can consist of printable ASCII characters.
        example: 6@5JG5hG
        maxLength: 72
        minLength: 8
        type: string
      newPassword:
        description: |-
          Password must contain at least one uppercase letter, one lowercase letter, one digit, and one special symbol.
          It can consist of printable ASCII characters.
        example: 6@5JG5hG
        maxLength: 72
        minLength: 8
        type: string
      oldPassword:
        description: |-
          Password must contain at least one uppercase letter, one lowercase letter, one digit, and one special symbol.
          It can consist of printable ASCII characters.
        example: 8p6R*R{3
        maxLength: 72
        minLength: 8
        type: string
    required:
    - confirmPassword
    - newPassword
    - oldPassword
    type: object
  dto.UserCreateDto:
    properties:
      password:
        description: |-
          Password must contain at least one uppercase letter, one lowercase letter, one digit, and one special symbol.
          It can consist of printable ASCII characters.
        example: 8p6R*R{3
        maxLength: 72
        minLength: 8
        type: string
      roleId:
        type: integer
      username:
        description: |-
          Username must start with an alphabetical character.
          It can consist of ASCII alphanumeric characters and the following symbols: _.-
        example: Username
        type: string
    required:
    - password
    - roleId
    - username
    type: object
  dto.UserUpdateDto:
    properties:
      active:
        type: boolean
      birthDate:
        description: Date only
        format: date
        type: string
      departments:
        items:
          type: integer
        type: array
      email:
        description: E-mail string.
        example: mail@mail.com
        type: string
      info:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
      name:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
      patronymic:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
      phone:
        description: E.164 phone number string.
        example: "+79876543210"
        type: string
      profession:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
      roleId:
        type: integer
      services:
        items:
          type: integer
        type: array
      sex:
        description: Alphabetic characters only (Russian and English).
        type: string
      surname:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Department:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      slug:
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Service:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      departments:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      active:
        type: boolean
      birthDate:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      departments:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      email:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      profession:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      sex:
        type: string
      slug:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vic.pozdeev@gmail.com
    name: Victor Pozdeev
  description: This is API specification for vet-clinic API server.
  license:
    name: MIT
    url: https://opensource.org/licenses/mit-license.php
  title: Vet clinic API
  version: v0.1.0
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login using username, e-mail, or phone along with the password.
      parameters:
      - description: Login and Password for logged-in.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to the authentication.
          headers:
            Cookie:
              description: Authorization
              type: string
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      summary: Login with credentials.
      tags:
      - Users
  /logout:
    post:
      consumes:
      - application/json
      description: Logout the user by invalidating the current session.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out.
        "401":
          description: Failed to the authentication.
      summary: Logout user.
      tags:
      - Users
...
  /profile:
    get:
      consumes:
      - application/json
      description: Retrieves and returns the user's own profile based on the session.
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get user's profile.
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates and returns the user's own profile based on the session.
      parameters:
      - description: User data for update.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Update user's profile.
      tags:
      - Users
  /profile/password:
    put:
      consumes:
      - application/json
      description: Updates the user's password based on the provided data.
      parameters:
      - description: Old password and new password for update.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Update user's password.
      tags:
      - Users
...
  /users:
    get:
      consumes:
      - application/json
      description: Returns the list of users.
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a user list.
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: A new user data for creating.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Create a new user. Required user''s role: Owner'
      tags:
      - Users
  /users/{id_or_slug}:
    get:
      consumes:
      - application/json
      description: Returns one record matched user's id or slug.
      parameters:
      - description: User ID
        in: path
        name: id_or_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a user.
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the existing user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Delete the existing user. Required user''s role: Superuser'
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the existing user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data for update.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Update the existing user. Required user''s role: Owner'
      tags:
      - Users
...
schemes:
- http
swagger: "2.0"